Feature/tsn 0019 like share service

+++ src/user_info_service/api/userInfo.ts
@@ -5,6 +5,7 @@ import { parse as uuidParse } from 'uuid';
 import { RequestWithPayload } from "../../auth_service/utils";
 import { utils } from "../../auth_service/utils/jwtUtils";
 import SQL_DB from "../../database";
+import { EMAIL, FIRST_NAME, MIDDLE_NAME, LAST_NAME, PROFILE_PIC_URL, JOINED_AT, UUID, ID, CREATED_AT, TITLE, SUMMARY, THUMBNAIL, TYPE, AUTHOR_ID, PUBLISH_STATUS, BOOKMARK_POST_ID, USER_UUID, FULL_STORY, POST_ID, FOLLOWER_ID, FOLLOWEE_ID } from "../../database/fields";
 
 const userInfoRouter: Router = Router();
 
@@ -12,26 +13,7 @@ const userInfoRouter: Router = Router();
 // 2. get user published posts
 // 3. get user unpublished posts
 // 4. get user bookmarks
-const EMAIL = "email",
-    FIRST_NAME ="first_name",
-    LAST_NAME = "last_name",
-    MIDDLE_NAME = "middle_name",
-    PROFILE_PIC = "profile_pic_url",
-    JOINED_AT = "created_at",
-    UUID = "uuid",
-    AUTHOR_ID = 'author_id',
-    USER_UUID = 'user_uuid',
-    CREATED_AT = 'created_at',
-    ID='id',
-    TITLE= 'title',
-    THUMBNAIL='thumbnail',
-    TYPE='type',
-    FULL_STORY_ID='full_story_id',
-    FULL_STORY = 'full_story',
-    PUBLISH_STATUS='publish_status',
-    SUMMARY='summary',
-    BOOKMARK_POST_ID='bookmark_post_id',
-    QUERY_SIZE = 50;
+const QUERY_SIZE = 50;
 
 type User = {email:string, id:string, iat:number|Date|string, exp:number|Date|string, aud:string, iss: string};
 
@@ -42,7 +24,7 @@ userInfoRouter.get('/personal-info/:userId', async (req_: Request, res: Response
             throw new createHttpError[404];
         } else {
             const db = new SQL_DB();
-            const response = await db.exec(db.TYPES.SELECT, `SELECT ${EMAIL}, ${FIRST_NAME}, ${MIDDLE_NAME}, ${LAST_NAME}, ${PROFILE_PIC}, ${JOINED_AT} FROM user WHERE ${UUID}=?`,
+            const response = await db.exec(db.TYPES.SELECT, `SELECT ${EMAIL}, ${FIRST_NAME}, ${MIDDLE_NAME}, ${LAST_NAME}, ${PROFILE_PIC_URL}, ${JOINED_AT} FROM user WHERE ${UUID}=?`,
             [Buffer.from(uuidParse(userID))]);
             if (response?.[0]?.[0]?.[EMAIL]) {
                 res.status(200).json({
@@ -51,7 +33,7 @@ userInfoRouter.get('/personal-info/:userId', async (req_: Request, res: Response
                         firstName: response[0][0][FIRST_NAME],
                         lastName: response[0][0][LAST_NAME],
                         middleName: response[0][0][MIDDLE_NAME],
-                        profilePicUrl: response[0][0][PROFILE_PIC],
+                        profilePicUrl: response[0][0][PROFILE_PIC_URL],
                         createdAt: response[0][0][JOINED_AT],
                     }
                 });
@@ -163,7 +145,7 @@ userInfoRouter.post('/bookmarks', utils.verifyAccessToken, async (req_: Request,
                 `WITH CTE AS (SELECT bookmark.${ID}, 
                 ROW_NUMBER() OVER(ORDER BY bookmark.${CREATED_AT} DESC) - 1 AS dataIndex, 
                 COUNT(*) OVER() AS totalCount, 
-                ${TITLE}, ${SUMMARY}, ${THUMBNAIL}, ${TYPE}, post.${CREATED_AT} AS createdAt
+                ${TITLE}, ${SUMMARY}, ${THUMBNAIL}, ${TYPE}, post.${CREATED_AT} AS createdAt, post.${AUTHOR_ID} AS authorID
                 FROM bookmark INNER JOIN post
                 ON bookmark.${BOOKMARK_POST_ID} = post.${ID}
                 WHERE ${USER_UUID} = ?
@@ -205,6 +187,7 @@ userInfoRouter.post('/drafts', utils.verifyAccessToken, async (req_: Request, re
                 COUNT(*) OVER() AS totalCount, ${FULL_STORY} AS fullStory, ${CREATED_AT} AS createdAt
                 FROM user_to_post 
                 WHERE ${AUTHOR_ID} = ?
+                AND ${POST_ID} IS NULL
                 ORDER BY ${CREATED_AT} DESC)
                 SELECT * FROM CTE WHERE dataIndex >= ? AND dataIndex < ?`,
             [Buffer.from(uuidParse(userID)), rangeStart, _rangeEnd]);
@@ -282,7 +265,7 @@ userInfoRouter.delete('/draft/delete', utils.verifyAccessToken, async(req_: Requ
     }
 });
 
-userInfoRouter.delete('/bookmark/delete', utils.verifyAccessToken, async(req_: Request, res: Response, next:NextFunction) => {
+userInfoRouter.post('/bookmarked', utils.verifyAccessToken, async(req_: Request, res: Response, next: NextFunction) => {
     try {
         const req = req_ as RequestWithPayload;
         const userID = req.body.userId || null;
@@ -293,12 +276,72 @@ userInfoRouter.delete('/bookmark/delete', utils.verifyAccessToken, async(req_: R
             throw new createHttpError[404];
         } else {
             const db = new SQL_DB();
+            const response = await db.exec(db.TYPES.SELECT, 
+                `SELECT ${ID} FROM bookmark WHERE ${BOOKMARK_POST_ID}=? AND ${USER_UUID} = ?`, 
+                [bookmarkID, Buffer.from(uuidParse(userID))]
+            );
+            if (response[0] && response[0].affectedRows === 1) {
+                res.status(200).json({
+                    success: true
+                });
+            } else {
+                res.status(200).json({
+                    success: false
+                });
+            }
+        }
+
+    }catch(error) {
+        next(error);
+    } 
+});
+userInfoRouter.put('/bookmark', utils.verifyAccessToken, async(req_: Request, res: Response, next:NextFunction) => {
+    try {
+        const req = req_ as RequestWithPayload;
+        const userID = req.body.userId || null;
+        const postID = req.body.postId || null;
+        if (!userID || userID == "") {
+            throw new createHttpError.BadRequest("User ID not found in query param");
+        } else if (!postID && !Number.isInteger(postID)) {
+            throw new createHttpError[404];
+        } else {
+            const db = new SQL_DB();
+            const response = await db.exec(db.TYPES.INSERT, "INSERT INTO `bookmark` SET ?", {
+                [USER_UUID]: Buffer.from(uuidParse(userID)),
+                [BOOKMARK_POST_ID]: postID
+            });
+            if (response[0] && response[0].affectedRows == 1) {
+                res.status(200).json({
+                    success: true
+                });
+            } else {
+                res.status(500).json({
+                    success: false
+                });
+            }
+        }
+    }catch(error) {
+        next(error);
+    } 
+});
+
+userInfoRouter.delete('/delete-bookmark', utils.verifyAccessToken, async(req_: Request, res: Response, next:NextFunction) => {
+    try {
+        const req = req_ as RequestWithPayload;
+        const userID = req.body.userId || null;
+        const postID = req.body.postId || null;
+        if (!userID || userID == "") {
+            throw new createHttpError.BadRequest("User ID not found in query param");
+        } else if (!postID && !Number.isInteger(postID)) {
+            throw new createHttpError[404];
+        } else {
+            const db = new SQL_DB();
             const response = await db.exec(db.TYPES.DELETE, 
                 `DELETE FROM bookmark 
-                WHERE ${ID}=?
+                WHERE ${BOOKMARK_POST_ID}=?
                 AND ${USER_UUID}=?`,
-            [bookmarkID, Buffer.from(uuidParse(userID))]);
-            if (response.affectedRows == 1) {
+            [postID, Buffer.from(uuidParse(userID))]);
+            if (response[0] && response[0].affectedRows == 1) {
                 res.status(200).json({
                     success: true
                 });
@@ -314,4 +357,100 @@ userInfoRouter.delete('/bookmark/delete', utils.verifyAccessToken, async(req_: R
     } 
 });
 
+userInfoRouter.post('/following', utils.verifyAccessToken, async(req_: Request, res: Response, next: NextFunction) => {
+    try {
+        const req = req_ as RequestWithPayload;
+        const followerID = req.body.followerId || null;
+        const followingID = req.body.followingId || null;
+        if (!followerID || !followingID) {
+            throw new createHttpError.BadRequest("Follower or followee Id is defined");
+        } else if(followerID === followingID) {
+            throw new createHttpError.BadRequest("Follower or followee Id is same");
+        } else {
+            const db = new SQL_DB();
+            const response = await db.exec(db.TYPES.SELECT, 
+                `SELECT ${ID} FROM follow WHERE ${FOLLOWER_ID}=? AND ${FOLLOWEE_ID} = ?`, 
+                [Buffer.from(uuidParse(followerID)), Buffer.from(uuidParse(followingID))]
+            );
+            if (response[0] && response[0].affectedRows === 1) {
+                res.status(200).json({
+                    success: true
+                });
+            } else {
+                res.status(200).json({
+                    success: false
+                });
+            }
+        }
+
+    }catch(error) {
+        next(error);
+    } 
+});
+
+userInfoRouter.put('/follow-request', utils.verifyAccessToken, async(req_: Request, res: Response, next: NextFunction) => {
+    try {
+        const req = req_ as RequestWithPayload;
+        const followerID = req.body.followerId || null;
+        const followingID = req.body.followingId || null;
+        if (!followerID || !followingID) {
+            throw new createHttpError.BadRequest("Follower or followee Id is defined");
+        } else if(followerID === followingID) {
+            throw new createHttpError.BadRequest("Follower or followee Id is same");
+        } else {
+            const db = new SQL_DB();
+            const response = await db.exec(db.TYPES.INSERT, 
+                "INSERT INTO `follow` SET ?", {
+                    [FOLLOWEE_ID]: Buffer.from(uuidParse(followingID)),
+                    [FOLLOWER_ID]: Buffer.from(uuidParse(followerID)),
+                } 
+            );
+            if (response[0] && response[0].affectedRows === 1) {
+                res.status(200).json({
+                    success: true
+                });
+            } else {
+                res.status(500).json({
+                    success: false
+                });
+            }
+        }
+
+    }catch(error) {
+        next(error);
+    } 
+});
+
+userInfoRouter.delete('/unfollow-request', utils.verifyAccessToken, async(req_: Request, res: Response, next: NextFunction) => {
+    try {
+        const req = req_ as RequestWithPayload;
+        const followerID = req.body.followerId || null;
+        const followingID = req.body.followingId || null;
+        if (!followerID || !followingID) {
+            throw new createHttpError.BadRequest("Follower or followee Id is defined");
+        } else if(followerID === followingID) {
+            throw new createHttpError.BadRequest("Follower or followee Id is same");
+        } else {
+            const db = new SQL_DB();
+            const response = await db.exec(db.TYPES.DELETE, 
+                `DELETE FROM follow WHERE ${FOLLOWER_ID}=? AND ${FOLLOWEE_ID}=?`, [
+                    Buffer.from(uuidParse(followerID)),
+                    Buffer.from(uuidParse(followingID))
+                ] 
+            );
+            if (response[0] && response[0].affectedRows === 1) {
+                res.status(200).json({
+                    success: true
+                });
+            } else {
+                res.status(500).json({
+                    success: false
+                });
+            }
+        }
+
+    }catch(error) {
+        next(error);
+    } 
+});
 export default userInfoRouter;
